Refactor marketing components and structure for improved organization and clarity. Moved landing page content to a new directory, added about and explore pages with respective content components, and established a marketing layout. Enhanced navigation and footer components for better user experience.
