+++ src/app/(marketing)/explore/[id]/page.tsx
