+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { LandingPageContent } from "./landing/landing-page-content";
+import { LandingPageContent } from "./(marketing)/landing/landing-page-content";
 
 export default function HomePage() {
   return <LandingPageContent />;
