+++ src/components/landing/organisms/how-it-works.tsx
