+++ src/components/landing/organisms/investment-opportunities.tsx
